@@ -1,15 +1,19 @@
 package org.zxp.ConcurrentLatch.demo.service;
 
-import org.zxp.ConcurrentLatch.LatchTaskName;
 import org.zxp.ConcurrentLatch.LatchThread;
 import org.zxp.ConcurrentLatch.demo.dto.PlatformDto;
 
-//@LatchTaskName("platform")
-public class PlatformLatch implements LatchThread {
+/**
+ * @program: ConcurrentLatch
+ * @description: 这个LatchThread演示无入参有返回值
+ * @author: X-Pacific zhang
+ * @create: 2019-06-12 10:09
+ **/
+public class PlatformLatch implements LatchThread<Void ,PlatformDto> {
 
     @Override
-    public PlatformDto handle() {
-        System.out.println("我是AAA");
+    public PlatformDto handle(Void v) {
+        System.out.println("我是PlatformLatch");
         try {
             Thread.sleep(1000);
         } catch (InterruptedException e) {
@@ -21,4 +25,5 @@ public class PlatformLatch implements LatchThread {
         dto.setPolicyNo("000000000001");
         return dto;
     }
+
 }
